feat(register): do not allow two nodes with same token
In some scenarios, same application and device registers against the
same appserver with the same token but different app user.  This patch
finds any existing registration with same token and deletes it before
writing the new one.  If we don't do so, it might happen that a user
receives notifications originally addressed to others.